Code refactor after rebase
