+++ infraohjelmointi_api/serializers/ProjectClassSerializer.py
@@ -1,8 +1,8 @@
 from infraohjelmointi_api.models import ProjectClass
-from infraohjelmointi_api.serializers import BaseMeta
+from infraohjelmointi_api.serializers import BaseMeta, FinancialSumSerializer
 from rest_framework import serializers
 
 
-class ProjectClassSerializer(serializers.ModelSerializer):
+class ProjectClassSerializer(FinancialSumSerializer, serializers.ModelSerializer):
     class Meta(BaseMeta):
         model = ProjectClass
