+++ infraohjelmointi_api/serializers/ProjectLocationSerializer.py
@@ -1,8 +1,8 @@
 from infraohjelmointi_api.models import ProjectLocation
-from infraohjelmointi_api.serializers import BaseMeta
+from infraohjelmointi_api.serializers import BaseMeta, FinancialSumSerializer
 from rest_framework import serializers
 
 
-class ProjectLocationSerializer(serializers.ModelSerializer):
+class ProjectLocationSerializer(FinancialSumSerializer, serializers.ModelSerializer):
     class Meta(BaseMeta):
         model = ProjectLocation
