+++ infraohjelmointi_api/validators/ProjectFinancialValidators/LockedFieldsValidator.py
@@ -8,19 +8,17 @@ class LockedFieldsValidator:
         project = serializer.instance
         if hasattr(project, "lock"):
             lockedFields = [
-                "phase",
-                "planningStartYear",
-                "constructionEndYear",
-                "programmed",
-                "projectClass",
-                "projectLocation",
-                "siteId",
-                "realizedCost",
-                "budgetOverrunAmount",
-                "budgetForecast1CurrentYear",
-                "budgetForecast2CurrentYear",
-                "budgetForecast3CurrentYear",
-                "budgetForecast4CurrentYear",
+                "budgetProposalCurrentYearPlus0",
+                "budgetProposalCurrentYearPlus1",
+                "budgetProposalCurrentYearPlus2",
+                "preliminaryCurrentYearPlus3",
+                "preliminaryCurrentYearPlus4",
+                "preliminaryCurrentYearPlus5",
+                "preliminaryCurrentYearPlus6",
+                "preliminaryCurrentYearPlus7",
+                "preliminaryCurrentYearPlus8",
+                "preliminaryCurrentYearPlus9",
+                "preliminaryCurrentYearPlus10",
             ]
             for field in lockedFields:
                 if allFields.get(field, None) is not None:
