+++ infraohjelmointi_api/validators/ProjectFinancialValidators/__init__.py
@@ -1,0 +1,1 @@
+from .LockedFieldsValidator import LockedFieldsValidator
