+++ infraohjelmointi_api/validators/__init__.py
@@ -1,2 +1,3 @@
 from .ProjectGroupValidators import ProjectsFieldValidator
 from .ProjectValidators import EstPlanningEndValidator, EstPlanningStartValidator
+from .ProjectFinancialValidators import LockedFieldsValidator
